+++ web/apps/photos/src/services/watch.ts
@@ -11,13 +11,13 @@ import type {
     FolderWatch,
     FolderWatchSyncedFile,
 } from "@/base/types/ipc";
+import { UPLOAD_RESULT } from "@/gallery/services/upload";
 import type { Collection } from "@/media/collection";
 import { EncryptedEnteFile } from "@/media/file";
 import {
     getLocalFiles,
     groupFilesByCollectionID,
 } from "@/new/photos/services/files";
-import { UPLOAD_RESULT } from "@/new/photos/services/upload/types";
 import { ensureString } from "@/utils/ensure";
 import debounce from "debounce";
 import uploadManager, {
