+++ web/packages/gallery/services/ffmpeg/index.ts
@@ -2,16 +2,16 @@ import { ensureElectron } from "@/base/electron";
 import log from "@/base/log";
 import type { Electron } from "@/base/types/ipc";
 import {
+    toDataOrPathOrZipEntry,
+    type DesktopUploadItem,
+    type UploadItem,
+} from "@/gallery/services/upload";
+import {
     readConvertToMP4Done,
     readConvertToMP4Stream,
     writeConvertToMP4Stream,
 } from "@/gallery/utils/native-stream";
 import { parseMetadataDate, type ParsedMetadata } from "@/media/file-metadata";
-import {
-    toDataOrPathOrZipEntry,
-    type DesktopUploadItem,
-    type UploadItem,
-} from "@/new/photos/services/upload/types";
 import {
     ffmpegPathPlaceholder,
     inputPathPlaceholder,
